Move runtime script type definition to defineRuntime.ts
